+++ src/pages/campaign/point/AddCampaignPoint.tsx
@@ -1,4 +1,3 @@
-import { render } from "@testing-library/react";
 import {
   Button,
   Col,
@@ -12,21 +11,96 @@ import {
   TimePicker,
 } from "antd";
 import moment from "moment";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { BackIconButton } from "../../../components/button/BackButton";
 import { CardContainer } from "../../../components/card/CardContainer";
 import FooterPage from "../../../components/footer/FooterPage";
 import { CardHeader } from "../../../components/header/CardHearder";
 import Layouts from "../../../components/layout/Layout";
+import { CampaignDatasource } from "../../../datasource/CampaignDatasource";
+import { CreateCampaignEntiry } from "../../../entities/CampaignPointEntites";
 import { color } from "../../../resource";
 
+const _ = require("lodash");
+
 const AddCampaignPoint = () => {
+  const profile = JSON.parse(localStorage.getItem("profile") || "{  }");
   const navigate = useNavigate();
   const dateFormat = "DD/MM/YYYY";
   const [form] = Form.useForm();
 
   const [showModal, setShowModal] = useState(false);
+  const [create, setCreate] = useState<CreateCampaignEntiry>();
+  const [checkDup, setCheckDup] = useState(false);
+
+  useEffect(() => {}, [checkDup]);
+
+  const checkDupCampiagn = async () => {
+    const getForm = form.getFieldsValue();
+    let startDate = new Date(
+      moment(getForm.startDate).format("YYYY-MM-DD")
+    ).toISOString();
+    let endDate = new Date(
+      moment(getForm.endDate).format("YYYY-MM-DD")
+    ).toISOString();
+    let application = getForm.application;
+    let check = await CampaignDatasource.checkDupCampaign(
+      "POINT",
+      startDate,
+      endDate,
+      application,
+    ).then((res) => {
+      if (!res.success) {
+        setCheckDup(true);
+      } else {
+        setCheckDup(false);
+      }
+      return !res.success;
+    });
+    return check;
+  };
+  const createCampaign = (e: any) => {
+    const getForm = form.getFieldsValue();
+    const data: any = { ...create };
+    const condition: any = { ...create?.condition };
+    condition.num = 1;
+    condition.point = parseFloat(getForm.point);
+    condition.rai = parseFloat(getForm.rai);
+    condition.rewardId = null;
+
+    data.campaignName = getForm.campaignName;
+    data.campaignType = "POINT";
+    data.application = getForm.application;
+    data.status = getForm.status;
+    data.condition = [condition];
+    data.createBy = profile.firstname + " " + profile.lastname;
+    data.updateBy = profile.firstname + " " + profile.lastname;
+    data.startDate = new Date(
+      moment(getForm.startDate).format("YYYY-MM-DD") +
+        " " +
+        moment(getForm.startTime).format("HH:mm:ss")
+    ).toISOString();
+    data.endDate = new Date(
+      moment(getForm.endDate).format("YYYY-MM-DD") +
+        " " +
+        moment(getForm.endTime).format("HH:mm:ss")
+    ).toISOString();
+    setCreate(data);
+    CampaignDatasource.createCampaign(data).then((res) => {
+      if (res.success) {
+        window.location.href = "/IndexCampaignPoint";
+      } else {
+        if (res.userMessage === "dupplicate") {
+        } else {
+        }
+      }
+    });
+  };
+  const submit = async () => {
+    await form.validateFields();
+    setShowModal(!showModal);
+  };
 
   return (
     <>
@@ -41,6 +115,9 @@ const AddCampaignPoint = () => {
           <CardHeader textHeader="ข้อมูลแคมเปญคะแนน" />
           <Form style={{ padding: "32px" }} form={form}>
             <Col span={24}>
+              <label>
+                ชื่อแคมเปญคะแนน<span style={{ color: color.Error }}>*</span>
+              </label>
               <Form.Item
                 name="campaignName"
                 rules={[
@@ -50,134 +127,161 @@ const AddCampaignPoint = () => {
                   },
                 ]}
               >
-                <label>
-                  ชื่อแคมเปญคะแนน<span style={{ color: color.Error }}>*</span>
-                </label>
-                <Input placeholder="กรอกชื่อแคมเปญคะแนน " />
+                <Input placeholder="กรอกชื่อแคมเปญคะแนน" autoComplete="off" />
               </Form.Item>
             </Col>
             <Row>
               <Col span={7}>
-                <Form.Item
-                  name="campaignName"
-                  rules={[
-                    {
-                      required: true,
-                      message: "กรุณากรอกชื่อแคมเปญคะแนน!",
-                    },
-                  ]}
-                >
-                  <label>
-                    วันเริ่มต้น<span style={{ color: color.Error }}>*</span>
-                  </label>
-                  <div className="d-flex">
-                    <Form.Item
-                      name="DateStart"
-                      rules={[
-                        {
-                          required: true,
-                          message: "กรุณากรอกวันที่!",
+                <label>
+                  วันเริ่มต้น<span style={{ color: color.Error }}>*</span>
+                </label>
+                <div className="d-flex">
+                  <Form.Item
+                    dependencies={[
+                      "endDate",
+                      "application",
+                      "startTime",
+                      "endTime",
+                    ]}
+                    name="startDate"
+                    rules={[
+                      {
+                        required: true,
+                        message: "กรุณากรอกวันที่!",
+                      },
+                      {
+                        validator: (rules, value) => {
+                          return new Promise(async (resolve, reject) => {
+                            if (await checkDupCampiagn()) {
+                              reject("");
+                            } else {
+                              resolve(true);
+                            }
+                          });
                         },
-                      ]}
-                    >
-                      <DatePicker
-                        placeholder="เลือกวันที่"
-                        format={dateFormat}
-                      />
-                    </Form.Item>
-                    <Form.Item
-                      name="TimeStart"
-                      initialValue={moment("00:00", "HH:mm")}
-                    >
-                      <TimePicker
-                        format={"HH:mm"}
-                        className="ms-3"
-                        placeholder="เลือกเวลา"
-                        defaultValue={moment("00:00", "HH:mm")}
-                        allowClear={false}
-                      />
-                    </Form.Item>
-                  </div>
-                </Form.Item>
+                      },
+                    ]}
+                  >
+                    <DatePicker
+                      placeholder="เลือกวันที่"
+                      format={dateFormat}
+                      onChange={checkDupCampiagn}
+                    />
+                  </Form.Item>
+                  <Form.Item
+                    name="startTime"
+                    initialValue={moment("00:00", "HH:mm")}
+                  >
+                    <TimePicker
+                      format={"HH:mm"}
+                      className="ms-3"
+                      placeholder="เลือกเวลา"
+                      defaultValue={moment("00:00", "HH:mm")}
+                      allowClear={false}
+                    />
+                  </Form.Item>
+                </div>
               </Col>
               <Col span={12}>
-                <Form.Item
-                  name="campaignName"
-                  rules={[
-                    {
-                      required: true,
-                      message: "กรุณากรอกชื่อแคมเปญคะแนน!",
-                    },
-                  ]}
-                >
-                  <label>
-                    วันสิ้นสุด<span style={{ color: color.Error }}>*</span>
-                  </label>
-                  <Col className="d-flex">
-                    <Form.Item
-                      name="DateEnd"
-                      rules={[
-                        {
-                          required: true,
-                          message: "กรุณากรอกวันที่!",
+                <label>
+                  วันสิ้นสุด<span style={{ color: color.Error }}>*</span>
+                </label>
+                <Col className="d-flex">
+                  <Form.Item
+                    dependencies={[
+                      "endDate",
+                      "application",
+                      "startTime",
+                      "endTime",
+                    ]}
+                    name="endDate"
+                    rules={[
+                      {
+                        required: true,
+                        message: "กรุณากรอกวันที่!",
+                      },
+                      {
+                        validator: (rules, value) => {
+                          return new Promise(async (resolve, reject) => {
+                            if (await checkDupCampiagn()) {
+                              reject("");
+                            } else {
+                              resolve(true);
+                            }
+                          });
                         },
-                      ]}
-                    >
-                      <DatePicker
-                        placeholder="เลือกวันที่"
-                        format={dateFormat}
-                      />
-                    </Form.Item>
-                    <Form.Item
-                      name="TimeEnd"
-                      initialValue={moment("00:00", "HH:mm")}
-                    >
-                      <TimePicker
-                        format={"HH:mm"}
-                        className="ms-3"
-                        placeholder="เลือกเวลา"
-                        defaultValue={moment("00:00", "HH:mm")}
-                        allowClear={false}
-                      />
-                    </Form.Item>
-                  </Col>
-                </Form.Item>
+                      },
+                    ]}
+                  >
+                    <DatePicker
+                      placeholder="เลือกวันที่"
+                      format={dateFormat}
+                      onChange={checkDupCampiagn}
+                    />
+                  </Form.Item>
+                  <Form.Item
+                    name="endTime"
+                    initialValue={moment("23:59", "HH:mm")}
+                  >
+                    <TimePicker
+                      format={"HH:mm"}
+                      className="ms-3"
+                      placeholder="เลือกเวลา"
+                      defaultValue={moment("00:00", "HH:mm")}
+                      allowClear={false}
+                    />
+                  </Form.Item>
+                </Col>
               </Col>
             </Row>
+            {checkDup && (
+              <p style={{ color: color.Error }}>
+                กรุณาเปลี่ยนแปลงช่วงเวลา “วันเริ่มต้น” หรือ “วันสิ้นสุด”
+                เนื่องจากซ้ำกับช่วงเวลาของแคมเปญอื่นที่สร้างไว้ก่อนหน้า
+              </p>
+            )}
             <Row gutter={8} justify={"start"}>
               <Col span={7}>
+                <label>
+                  คะแนนที่ได้รับ<span style={{ color: color.Error }}>*</span>
+                </label>
                 <Form.Item
-                  name="campaignName"
+                  name="point"
                   rules={[
                     {
                       required: true,
-                      message: "กรุณากรอกชื่อแคมเปญคะแนน!",
+                      message: "กรุณากรอกจำนวนคะแนน!",
                     },
                   ]}
                 >
-                  <label>
-                    คะแนนที่ได้รับ<span style={{ color: color.Error }}>*</span>
-                  </label>
-                  <Input placeholder="กรอกคะแนนที่ได้รับ" suffix="คะแนน" />
+                  <Input
+                    placeholder="กรอกคะแนนที่ได้รับ"
+                    suffix="คะแนน"
+                    autoComplete="off"
+                  />
                 </Form.Item>
               </Col>
               <Col>
                 <label style={{ paddingTop: "25px" }}> : </label>
               </Col>
               <Col span={7}>
+                <label>
+                  จำนวนไร่ <span style={{ color: color.Error }}>*</span>
+                </label>
                 <Form.Item
-                  name="campaignName"
+                  name="rai"
                   rules={[
                     {
                       required: true,
-                      message: "กรุณากรอกชื่อแคมเปญคะแนน!",
+                      message: "กรุณากรอกจำนวนไร่!",
                     },
                   ]}
                 >
-                  <label>
-                    จำนวนไร่ <span style={{ color: color.Error }}>*</span>
-                  </label>
-                  <Input placeholder="กรอกจำนวนไร่ " suffix="ไร่" />
+                  <Input
+                    placeholder="กรอกจำนวนไร่ "
+                    suffix="ไร่"
+                    autoComplete="off"
+                  />
                 </Form.Item>
               </Col>
             </Row>
@@ -194,9 +298,12 @@ const AddCampaignPoint = () => {
                   },
                 ]}
               >
-                <Radio.Group className="d-flex flex-column">
-                  <Radio value={"FARMER"}>Farmer Application</Radio>
-                  <Radio value={"DRONER"}>Droner Application</Radio>
+                <Radio.Group
+                  className="d-flex flex-column"
+                  onChange={checkDupCampiagn}
+                >
+                  <Radio value={"FARMER"}>Farmer</Radio>
+                  <Radio value={"DRONER"}>Droner</Radio>
                 </Radio.Group>
               </Form.Item>
             </Col>
@@ -205,7 +312,7 @@ const AddCampaignPoint = () => {
                 สถานะ <span style={{ color: color.Error }}>*</span>
               </label>
               <Form.Item
-                name="application"
+                name="status"
                 rules={[
                   {
                     required: true,
@@ -215,7 +322,7 @@ const AddCampaignPoint = () => {
               >
                 <Radio.Group className="d-flex flex-column">
                   <Radio value={"ACTIVE"}>ใช้งาน</Radio>
-                  <Radio value={"DRAFT"}>รอเปิดใช้งาน</Radio>
+                  <Radio value={"DRAFTING"}>รอเปิดใช้งาน</Radio>
                 </Radio.Group>
               </Form.Item>
             </Col>
@@ -224,7 +331,7 @@ const AddCampaignPoint = () => {
         <FooterPage
           onClickBack={() => navigate(-1)}
           styleFooter={{ padding: "6px" }}
-          onClickSave={() => setShowModal(!showModal)}
+          onClickSave={() => submit()}
         />
       </Layouts>
       {showModal && (
@@ -270,6 +377,7 @@ const AddCampaignPoint = () => {
                 backgroundColor: color.Success,
                 color: color.White,
               }}
+              onClick={createCampaign}
             >
               ยืนยัน
             </Button>